opp-api/requirements.txt
aiofiles==23.2.1
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
astroid==3.0.1
asttokens==2.4.0
backcall==0.2.0
bcrypt==4.0.1
black==23.10.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
decorator==5.1.1
dill==0.3.7
ecdsa==0.18.0
executing==2.0.0
fastapi==0.103.2
h11==0.14.0
httptools==0.6.0
idna==3.4
ipython==8.16.1
isort==5.12.0
jedi==0.19.1
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
parso==0.8.3
passlib==1.7.4
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pygments==2.16.1
pylint==3.0.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
stack-data==0.6.3
starlette==0.27.0
tomlkit==0.12.1
traitlets==5.11.2
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
wcwidth==0.2.8
websockets==11.0.3
requests==2.31.0
jose~=1.0.0